calculation error prevention
