+++ components/mainFoodPage/styles/calcErrorPrevent-popup.module.css
@@ -1,21 +1,19 @@
 /* Clear Calculator Error Prevention STYLES */
 .offWhiteBG{
-    /* display: hidden; */
     z-index: 10000;
-    display: flex;
-    justify-content: center;
-    align-items: center;
 
-    background-color: rgba(255, 255, 255, 0.5);
     position: fixed;
     top: 0;
     left: 0;
     width: 100%;
     height: 100%;
-    
-    cursor:default;
+    background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent black background */
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
 .errorPreventPopup{
+    /* z-index: 10000; */
 
     background: #F6EDE4;
 
@@ -53,4 +51,19 @@
     border: 1px solid rgba(0, 0, 0, 0.16);
     font-size: 35px;
     cursor: pointer;
+}
+
+.clearCalcButton {
+    height: 50px;
+    width: 95%;
+    padding: 10px;
+    box-sizing: border-box;
+    border-radius: 10px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 1px solid rgba(0, 0, 0, 0.16); /* Thin black border */
+    font-size: 25px;
+    cursor: pointer;
+    background-color: #c44d1a9c;
 }
